requirements.txt
Django==1.9.7
django-axes==1.7.0
django-custom-user==0.6
django-extensions==1.6.7
django-grappelli==2.8.1
django-guardian==1.4.4
djangorestframework==3.3.3
gunicorn==19.6.0
httpie==0.9.3
psycopg2==2.6.1
Pygments==2.1.3
requests==2.10.0
six==1.10.0
django-webpack-loader==0.3.3
